+++ backend/business/internal/repository/repository.go
@@ -13,20 +13,20 @@ type AuthRepo interface {
 // BusinessMemberRepo は事業者メンバーに関するデータアクセスメソッドを定義します。
 type BusinessMemberRepo interface {
 	GetByGoogleID(ctx context.Context, googleID string) (interface{}, error)
-	UpdateName(ctx context.Context, businessID int64, name string) error
-	UpdateIcon(ctx context.Context, businessID int64, icon []byte) error
-	Anonymize(ctx context.Context, businessID int64) error
+	UpdateName(ctx context.Context, businessID int, name string) error
+	UpdateIcon(ctx context.Context, businessID int, icon []byte) error
+	Anonymize(ctx context.Context, businessID int) error
 }
 
 // PostRepo は投稿に関するデータアクセスメソッドを定義します。
 type PostRepo interface {
-	ListByBusiness(ctx context.Context, businessID int64) (interface{}, error)
-	GetByID(ctx context.Context, postID int64) (interface{}, error)
-	Create(ctx context.Context, businessID int64, placeID int64, genreIDs []int64, payload interface{}) (int64, error)
-	SetGenres(ctx context.Context, postID int64, genreIDs []int64) error
+	ListByBusiness(ctx context.Context, businessID int) (interface{}, error)
+	GetByID(ctx context.Context, postID int) (interface{}, error)
+	Create(ctx context.Context, businessID int, placeID int, genreIDs []int, payload interface{}) (int, error)
+	SetGenres(ctx context.Context, postID int, genreIDs []int) error
 	// IncrementViewCount は投稿の閲覧数を1増やします
-	IncrementViewCount(ctx context.Context, postID int64) error
-	Anonymize(ctx context.Context, postID int64) error
+	IncrementViewCount(ctx context.Context, postID int) error
+	Anonymize(ctx context.Context, postID int) error
 	History(ctx context.Context, googleID string) (interface{}, error)
 }
 
@@ -48,30 +48,30 @@ type ContactRepo interface {
 
 // StatsRepo はダッシュボード統計に関するデータアクセスメソッドを定義します。
 type StatsRepo interface {
-	TotalPosts(ctx context.Context, businessID int64) (int64, error)
-	TotalReactions(ctx context.Context, businessID int64) (int64, error)
-	TotalViews(ctx context.Context, businessID int64) (int64, error)
-	EngagementStats(ctx context.Context, businessID int64) (int64, int64, int64, error)
+	TotalPosts(ctx context.Context, businessID int) (int, error)
+	TotalReactions(ctx context.Context, businessID int) (int, error)
+	TotalViews(ctx context.Context, businessID int) (int, error)
+	EngagementStats(ctx context.Context, businessID int) (int, int, int, error)
 }
 
 // PaymentRepo は支払いに関するデータアクセスメソッドを定義します。
 type PaymentRepo interface {
-	CreatePayment(ctx context.Context, businessID int64, amount int64, payFlag bool) (int64, error)
+	CreatePayment(ctx context.Context, businessID int, amount int, payFlag bool) (int, error)
 }
 
 // ReactionRepo はリアクションに関するデータアクセスメソッドを定義します。
 type ReactionRepo interface {
-	CreateUnique(ctx context.Context, userID string, postID int64) error // (userId, postId) の組み合わせが一意であることを保証する必要があります
-	CountByPostIDs(ctx context.Context, postIDs []int64) (map[int64]int64, error)
+	CreateUnique(ctx context.Context, userID string, postID int) error // (userId, postId) の組み合わせが一意であることを保証する必要があります
+	CountByPostIDs(ctx context.Context, postIDs []int) (map[int]int, error)
 }
 
 // PlaceRepo は場所に関するデータアクセスメソッドを定義します。
 type PlaceRepo interface {
-	GetByID(ctx context.Context, placeID int64) (interface{}, error)
+	GetByID(ctx context.Context, placeID int) (interface{}, error)
 }
 
 // GenreRepo はジャンルに関するデータアクセスメソッドを定義します。
 type GenreRepo interface {
-	GetByID(ctx context.Context, genreID int64) (interface{}, error)
-	ListByPostID(ctx context.Context, postID int64) ([]int64, error)
+	GetByID(ctx context.Context, genreID int) (interface{}, error)
+	ListByPostID(ctx context.Context, postID int) ([]int, error)
 }
